Move line string mesh building logic to `linestring.rs`
